_threadmodule: thread-safety fixes
 - Make rlock thread-safe
 - Use atomics to modify interp->num_threads

+++ Include/internal/pycore_interp.h
@@ -90,7 +90,7 @@ struct _is {
     PyObject *importlib;
 
     /* Used in Modules/_threadmodule.c. */
-    long num_threads;
+    Py_ssize_t num_threads;
     /* Support for runtime thread stack size tuning.
        A value of 0 means using the platform's default stack size
        or the size specified by the THREAD_STACK_SIZE macro. */
